wqbench: Calculate Aquatic Life Benchmarks

Download recent versions of the US EPA ECOTOX database. Clean, standardize and classify data so values are comparable. Use a SSD or deterministic method to determine the critical toxicity value and assessment factors to compute the aquatic life water quality benchmark for a compound.

See also

Author

Maintainer: Angeline Tillmanns Angeline.Tillmanns@gov.bc.ca

Authors:

Ayla Pearson ayla@poissonconsulting.ca (ORCID)

Other contributors:

Joe Thorley joe@poissonconsulting.ca (ORCID) [contributor]

Andy Teucher andy@andyteucher.ca (ORCID) [contributor]

Province of British Columbia [copyright holder]